Merged PR 211039: llvm symbolizer build cmake fix
LLVM build expects us to be building within the VS directory so when the build kicks off on a devdiv agent it gets confused.
This forces the build to compile the symbolizer stuff that depends on DIA being available. The other half of the build fix is in project files in msvc, so another PR will merge that part.