Merged PR 274845: initial cleanup work to get sanitizer_purge_allocator working
sanitizer_purge_allocator is designed to help fuzzing be efficient, you can clear out the allocator to start over without reloading the whole process. We didn't implement this before for our moveable manager, this commit adds that functionality.

Additionally, since this object is now needed across two files part of the fix is switching from using a single static instance in malloc_win to a singleton pattern, making a static instance in the malloc_moveable file and accessing it elsewhere with a static method, GetInstance.