Merged PR 281397: Tag ASAN unit tests as known-failures
I ran our ASAN unit test workflow and automatically tagged each test that failed in any combination with msvc-host. I experimented locally with trying to "refine" the bail condition, but there seemed to be some unintuitive gotchas that I do not think it's appropriate to spend overmuch time addressing, so I'm going for the big-hammer. As a middle-ground, I added a comment of the logs that originally contained this failing test.

See update 13 of https://devdiv.visualstudio.com/DevDiv/_git/msvc/pullrequest/275028 for a demonstration that this gets all the tests.