Merged PR 295500: Bugfix: Create GlobalAlloc and LocalAlloc asan option
A previous merge removed a check for windows_hook_rtl_allocators before installing the global and local alloc interceptors, this was fine because the checks were in the functions themselves. After some changes, the checks are now missing. Additionally, the asan option name doesn't make sense anymore since the RTL functions are just always on.

This fixes the issue (adding in the check) and changes the ASan option to a name that makes more sense from the user's perspective.

This change requires changing the option name in some of the unit tests, and removing it in some other ones where it is no longer needed.