Merged PR 298908: Fixing large_func_test.cc failure
The Clang version of ASan unit tests uses the -fno-omit-frame-pointer flag when compiling the tests to obtain an accurate stack trace output in ASan error dumps. This option is added to the `clangxx_asan` and `clang_asan` variables in `lit.cfg.py`. However, the MSVC substitution mapping for these variables does not include the equivalent /Oy- option. This causes a spurious failure in `large_func_test.cc` when compiled with an upcoming version of MSVC where frame pointer omission happens.

This change adds the /Oy- substitution mapping to the `clangxx_asan` and `clang_asan` variables in `msvc_main.py` to achieve parity with Clang.

### Testing
Ran MSVC ASAN unit tests locally.