Merged PR 301619: Fixing VirtualAlloc memory leak
ReturnMemoryPagesToOS was previously a stub function, memory that was mapped was never unmapped.

Fixing this issue revealed a second problem where the ASan allocator expects re-mapping to be automatic on *nix platforms. Since Windows requires VirtualAlloc VirtualFree to be explicit to map-unmap-remap pages, a change to the allocator was required.

This change checks if memory in the region was released and re-maps it once the region is used again.