Merged PR 304705: Make it possible to link ASan DLLs which statically link in the debug CRT
Make it possible to link ASan DLLs which statically link in the debug CRT.
We also update the testing harness to not `/force:multiple` which is what was hiding this issue previously.
I'm aware that I don't provide the data symbols that are defined in the debug CRT but I'm not currently sure how to handle them correctly and would rather allow the link the fail than do something completely broken.

Related work items: #1279215