Merged PR 313914: Hook memcpy to point to memmove
On `x86` `memcpy` and `memmove` have the exact same implementation, but have different definitions. Because of this calls to `memmove` can randomly get changed to calls to `memcpy` by the linker when `OPT:ICF` is enabled.

The Asan runtime already handles `amd64` having the exact same definition for both `memcpy` and `memmove` by simply overwriting the function pointer for `memcpy`. We handle this on `x86` by intercepting both `memcpy` and `memmove` to point to the `memmove` hook.

Related work items: #1288174, #1294972, #1308544, #1314137