Merged PR 326034: Pre-commit shadow memory for ranges that we know of ahead of time
Pre-commit shadow memory for ranges that we know of ahead of time.

The following synthetic test runs anywhere from 1x - 3x the speed of the previous implementation depending on the number of threads, size of the allocations the threads are making and the number of allocations. I've pasted the code with the numbers that are least likely to lead to OOM and provide the best speedup numbers on my machine. (All code profiled by VS)

```

void allocations() {
    char *mems[NUM_ALLOCATIONS];
    for (int i = 0; i < NUM_ALLOCATIONS; ++i) {
        mems[i] = (char*)malloc(ALLOCATION_SIZE);
        *mems[i] = 'c';
        *(mems[i] + ALLOCATION_SIZE - 1) = 'd';
    }

    for (int i = 0; i < NUM_ALLOCATIONS; ++i) {
        free(mems[i]);
    }
}

int main() {
    std::thread threads[NUM_THREADS];
    for (int i = 0; i < NUM_THREADS; ++i) {
        threads[i] = std::thread(allocations);
    }

    for (int i = 0; i < NUM_THREADS; ++i) {
        threads[i].join();
    }
}
```

Related work items: #1270477