Merged PR 340344: Changes for LibFuzzer tests to run on MSVC
The goal here is to bring up the fuzzer tests in the ASAN repo for #1244749.

This change involves edits to the test runner to support the fuzzer tests, editing some of the tests so that they can build with our compiler, and adding 'msvc-host' to the UNSUPPORTED list for tests that try something we don't support yet. There are a couple tests that seem to indicate something is broken:
* caller-callee.test
    * This test has never found the target for me in my runs. I don't know the reason - presumably some /fsanitize-coverage option we don't support gives the runtime enough information? Let me know if you have an idea about this one.
* merge.test, merge_two_step.test
    * The 'F.....' input appears to be cursed and will always be added to the corpus, even on re-runs with no changes (so, it will literally say "1 file added to the corpus", even if no files were visibly added because it just keeps overwriting the 'F.....' one).
* reduce_inputs.test
    * We keep the 7 files and don't reduce them.

Thoughts on opening bugs for these or leaving these as simply disabled?

I also noticed some instability with the test runner? Some tests run fine alone, but sometimes fail in the test runner. I also noticed that, particularly with standalone.test, the replacement logic sometimes will do the -c -o -> /c /Fo replacement, but sometimes chooses -o -> /Fe instead and I could not nail down a consistent reason - any thoughts?

Changes in the MSVC repo are here: !340348

Related work items: #1244749