Merged PR 341317: Allow linking to APISET-only umbrella libraries
Link to corresponding MSVC PR: !341323

This resolves #1360650.

When delivered to Windows, some projects are failing to build with the static ASAN runtimes because `LocalHandle` is not resolved during link time. `LocalHandle` is only provided by kernel32 and is not available in an APISET. For historical reasons, the umbrella libraries (onecore.lib/onecoreuap.lib) that Windows uses to build are APISET-only. Prior for inclusion in the SDK, the imports for kernel32 are explicitly re-added to those libraries for user convenience. When we build and test ASAN, we are using the publicly released Windows SDK, so don't see this same issue. In order to support Windows, we also need to support linking against the APISET-only versions of onecore.lib and onecoreuap.lib. Luckily, these are also provided in the Windows SDK with the _apiset suffix.

This change adds a test to ensure that we can link a simple program with onecore_apiset.lib and onecoreuap_apiset.lib, as well as changes to `LocalHandle` usage so that we do not actually reference the symbol and instead it is lazy-loaded.

Some additional cleanup changes that I found are:
1) `#if defined(ASAN_DYNAMIC)` doesn't do anything. ASAN_DYNAMIC is either defined to 1 or 0, and I believe in this case we want to hook the system allocators regardless of whether we are linking statically or dynamically.
2) The `LocalHandle` wrapper called `GlobalHandle` instead of `LocalHandle` when ASAN had not been initialized yet.

Related work items: #1360650