Merged PR 346097: Fix RtlHeap API Lock Order Inversion
MSVC PR: !346098

Currently, when -trace_malloc is set, fuzzer::MallocHook and fuzzer::FreeHook is added to the malloc/free hook list. Both of these take the Trace lock, so every new or free call will start with acquiring the trace lock. Currently, RtlAllocateHeap and friends take a lock on the heap handle prior to calling asan_malloc/asan_free, so some calls to asan_malloc/asan_free have this heap handle lock held, and some do not. RtlAllocateHeap is called many times during asan_malloc, so this situation where RtlAllocateHeap calls asan_malloc under a lock and most other asan_malloc calls do not gets us into this lock order inversion issue. To fix this, we need to ensure that no exclusive-access lock is taken before calling asan_malloc or asan_free.

To do this, I've renamed the spin mutex used to guard access to the AsanHeap data to the MemoryMapLock. The only data that requires synchronized access is the memory map and memory list, so this lock is now only taken when access to those structures is needed (via the unnamed handle returned by MemoryMapLockGuard). This minimizes the amount of time where we are holding a lock.

The other responsibility of the original spin mutex was to ensure the AsanHeap isn't deleted while another thread is still using it via a call to RtlDestroyHeap. Using a lock to provide this guarantee proved to have too many restrictions and deadlock opportunities, so instead this is guaranteed by reference counting the AsanHeap. Now, RtlDestroyHeap simply removes the AsanHeap from the AsanHeapMap and decrements the reference count. Once the reference count hits zero, the AsanHeap destructor takes care of RtlDestroyHeap's original responsibilities.

To help ensure no logic errors are made, access to AsanHeap data is restricted to just AsanHeapHandle, and a DebugChecks object is created for each call in CHK builds to assert invariants that cannot be structurally guaranteed.

In the first iteration of this fix, I removed reentrancy checks in RtlFreeHeap, however this caused intermittent test failures. The Low Fragmentation Heap is reentrant for internal allocations, but those internal allocations will not show up when walking the heap. In order to handle these allocations, I added a check for the HEAP_NO_CACHE_BLOCK flag when the heap signature indicates we are using the LFH and redirect the call back to the real Rtl implementation in these cases (this was discussed with the Win32 Heap team). Originally, we only handled reentrancy in RtlFreeHeap, but this change also fixes this same reentrancy issue in RtlAllocateHeap and RtlReAllocateHeap - I've added the relevant bugs/work items.

Unfortunately, checking for HEAP_NO_CACHE_BLOCK does not work for frees and there is not a way to manually find the internal blocks by walking the heap without doing it manually, which the Win32 Heap strongly disagrees with doing. In addition to the HEAP_NO_CACHE_BLOCK checks, reentrancy checks were re-added. A new reentrancy lock is held every tim...