Merged PR 374338: Fix x86 strchr interception issues
MSVC PR: !374341

ntdll!strchr on x86 was crashing with some code paths (ex: `strchr("what-ever", "-")`) with the new interception changes to take Win11 XFG hashes into account. In particular, it used to be hooked via the trampoline method, not via hotpatch method. There isn't enough space for the hotpatch method on x86, but the changes force it to be used because it's trying to allow an XFG hash to be present.

Look at line 998 in interception_win.cpp:
```
if (instruction_size < kShortJumpInstructionLength ||
      (!guaranteed_hotpatchable &&
      !FunctionHasPadding(old_func, kHotPatchHeaderLen))) {
    return false;
  }
```
FunctionHasPadding would fail on x86, but it was ignored because it was "guaranteed hotpatchable".

Curtis added the guaranteed hotpatchable idea in a previous PR (!368664) because XFG hashes make it always look like it shouldn't be hotpatchable (the space for the header is where the XFG hash goes). The idea was that all windows binaries are built with /hotpatch, so there must be enough space to apply the hotpatch, but in practice it doesn't appear like there's enough function padding on x86. The lack of space for the header only affected the x86 binaries.

This change is simply to fix the regressions and to introduce a test to ensure that random input won't cause calling any of the intercepted string APIs to crash. In the future, I'm not sure we should be calling `ZwSetInformationVirtualMemory` directly like this (edit: this API *is* [documented](https://docs.microsoft.com/en-us/windows-hardware/drivers/ddi/ntifs/nf-ntifs-zwsetinformationvirtualmemory), just not with these parameters), and we need a better way to detect whether there's enough space to hotpatch regardless of whether an XFG hash is present.

Related work items: #1463596, #1463598, #1463597