Merged PR 372055: Fix for deadlock on windows process terminate
MSVC PR !372143

**Edit:**
Uses custom `PEB_LDR_DATA` to extract `ShutdownInProgress` flag which is set during `RtlExitUserProcess` in order to not reacquire locks in the asan runtime on process termination.

**Originally:**
Added an interceptor for `NtTerminateProcess` to make locking mechanisms no-ops due to dead threads holding onto locks.
`NtTerminateProcess` is called twice from `RtlExitUserProcess`, but the first time is when we need to set the flag.

Related work items: #1371405