Merged PR 379856: Track allocations prior to ASAN initialization
Corresponding msvc pr: !379857

Adds logic to track allocations that happen prior to when asan begins tracking the allocations in order to prevent false positive reporting.

This is one approach that _simply_ tracks the allocations that happen prior to asan initialization by walking the heap. Before we call any of the functions from the dlls that manipulate memory, we check to see if that block of memory was allocated prior to asan initialization. If it was, we delegate the call to the original dll that contains the function that allocated the memory. Depending on if that call manipulates memory or not, asan will take ownership for subsequent manipulations. If the memory was not allocated prior to asan initialization, we continue down the asan runtime callstack as normal.

**Relevant changes:**
-
- **asan_malloc_win.cpp:** most of the changes take place here. The `RuntimeFunctions` class is basically a host of function pointers combined with static functions for overriding the normal windows dll memory manipulating functions. The class checks for whether or not a block of memory was allocated prior to asan init, and depending on the result delegates to the correct functions.
- **asan_internal.h:** added `AllocatedBeforeAsanInit` and `CaptureSystemHeapAllocations` functions
- **asan_rtl.cpp:** added a call to `CaptureSystemHeapAllocations` during startup
- **asan_win.cpp:** added `SystemAllocationMap` to track allocations prior to asan initialization as well as `CaptureSystemHeapAllocations` definition
- **interception_win.h/cpp:** added logic to override functions based specifically on dll name
- **valid_memory_operations_prior_to_init.cpp:** tests both `_setmaxstdio` realloc and user allocations prior to asan initialization

Related work items: #1337321