Merged PR 382180: Remove STL and Refactor GlobalAlloc/LocalAlloc Allocators
MSVC Repo PR: !382184

The original purpose of this work is to remove the STL dependency from the legacy interceptors for GlobalAlloc/LocalAlloc. While doing this, I wanted to ensure that I was not introducing any regressions so added a whole bunch of behavior tests. What I found was that our current implementation a) provides slightly different behavior from the actual GlobalAlloc/LocalAlloc implementation and b) does not provide ASAN reports for most of these errors, instead choosing to use CHECK assertion messages.

Most relevant changes are in asan_malloc_win_moveable.cpp. I've attempted to refactor to prioritize clarity, by trying to separate responsibilities and contain invariants to as small as a scope as possible. If anything is unclear, please let me know and I can add more comments to the source file.

As a part of this, I also:
    * Added EmplaceBack(), const Back(), and Empty() functions to the __sanitizer::Vector implementation.
    * Implemented asan_recalloc in AsanAllocator.

Related work items: #1458781