Merged PR 409481: IAT Overwrite protection
Added protection inside of sanitizer_common to prevent import address table overwrites from causing sanitizer failures.
**MSVC PR:** !412036

A runtime flag `iat_overwrite` was added to control the level of protection for the consumer. The default is for these checks to be in place and error if an overwrite is detected (i.e. `iat_overwrite=error`). When set to `iat_overwrite=ignore`, we will ignore the checks and call whatever the resolution of `::VirtualAlloc`/`::VirtualQuery`/`::VirtualProtect` ends up being. When set to `iat_overwrite=protect`, we will attempt to call the correct `kernel32.dll!Virtual*` function.