Merged PR 422630: Intercept AddVectoredExceptionHandler and RemoveVectoredExceptionHandler
**MSVC PR:** !424890

CoreCLR confirmed this behaves as they needed. This intercepts both `AddVectoredExceptionHandler` and `RemoveVectoredExceptionHandler`. The implementation is handwritten assembly that will save the address of both the `ShadowExceptionHandler` and the user's requested exception handler to add in order to first call the `ShadowExceptionHandler` before continuing into the user's exception handling code.

One thing this does **not** do is free the memory allocated for storing the newly created functions. If we were to attempt this, we could potentially free the function before the `Rtl*` functions are through using it, causing a crash. In the future, depending on how we handle the shadow memory AV's, we could potentially fix this by some type of reference counting. It's likely not to be much overhead since `RemoveVectoredExceptionHandler` is used much less frequently or only on exit.

Related work items: #1595370