Merged PR 435149: Fix mixed_link_configuration failing test (/MT exe, /MD DLL)
MSVC PR: !435151

This change removes the CHECK from the interception call for `AddVectoredExceptionHandler`. We haven't typically required the intercept succeed and this has allowed scenarios where multiple of instances of ASAN are running at the same time (which is what happens in the /MT host and /MD DLL scenario - the DLL intercepts first, then the /MT exe will fail to override everything). This scenario isn't how we want this to work but happens to now so ideally we don't regress - we will be supporting this officially following work to allow mixed configurations.

Related work items: #1673597