Merged PR 441989: Use real Rtl* functions if memory is mapped.
**MSVC PR:** !441990

(Not sure how CI will interact with the test as far as IPC goes.)

This is the fix for the Word logout issue.

####How the issue happened:

`dcomp`, which is a component that creates shared memory and writes to it for `dwm` to read from, was the culprit. It has some internal heap management logic that will pass a relative virtual address offset to `dwm`, which is expected to be inside of a heap that is created inside of a mapped memory buffer. Essentially, `RtlCreateHeap` using mapped memory.

ASAN's `__asan_wrap_RtlAllocateHeap` effectively ignores the base address argument to `RtlAllocateHeap`, and allocates the memory inside of its own allocator, which is outside of the internal heap range mentioned above. This leads to problems reading the memory on the receiving side -- the offset they receive is not inside of the mapped memory they have loaded, and a fastfail causes the `dwm` crash, then subsequent logout because that's what Windows does.

####What the fix does:
The fix is to just ignore mapped memory. Since we don't allocate inside of that buffer anyways, we can't guarantee the same behavior that happens when the program isn't instrumented with ASAN.

####What we should do later?
I haven't written a follow up story yet, but I think we should discuss this sometime in the future. Maybe we can hook `RtlCreateHeap` and utilize the [HeapBase](https://learn.microsoft.com/en-us/windows-hardware/drivers/ddi/ntifs/nf-ntifs-rtlcreateheap#:~:text=%5Bin%2C%20optional%5D%20HeapBase) parameter..? I don't think that fully solves the problem of RVA offsets being passed between processes, though :)

Related work items: #1695679