Merged PR 450782: Changing `windows_hook_legacy_allocators` to default to true
Realized I only opened the **MSVC pr**... : !449655

The latest commit of this is just clang-format. I think the update to llvm14 changed the clang-format file, so it appears there are a lot more changes.

The change here is to make `windows_hook_legacy_allocators` default to true. The problem XStore was seeing was that for an absurdly large amount of threads, their test were taking orders of magnitude longer. All threads were trying to access or allocate on the same heap, which means each thread is waiting on the `HeapLock`.

In 17.2, the code path taken during the test would've been to allow the allocation to take place in the asan allocator. After the global / local alloc rework and depending on the flags passed into `RtlAllocateHeap`, the allocation _may_ be handled by `REAL(RtlAllocateHeap)`, which results in both the reentrancy lock being taken as well as the `HeapLock`.

After some testing, it looks like we can safely change the default option to be true, while maintaining current behavior for `windows_hook_legacy_allocators=false` if any customer is seeing issues. I couldn't find out why some tests were marked `xfail`, but I was able to turn them all on with no issues (both locally and in the CI builds).

This also needs an associated documentation update to mark the option as defaulting to true now.

Related work items: #1694097