Merged PR 475804: Fix x86 hang when _except_handler4 is intercepted twice
MSVC PR: !475486

This is the fix for the issue caught in MSVC-ASAN-UnitTests-Rolling (see !474849).

When the `_except_handler4` interceptor runs twice, it ends up with `real_except_handler` pointing to `intercept_except_handler4`, causing any use of `_except_handler4` to result in a hang (not a stack overflow since the generated code elides the frame). The CRT initializer always runs, but the TLS initializer only runs when there is thread local data to initialize, hence why we only see this causing an issue on one fuzzer test case (minimize_two_crashes.test) that happens to use both TLS and `_except_handler4`.

`_except_handler4` is the only affected function because it is the only function that uses this "by-address" interception call. The function definition must exist within the calling module in order to have the correct security cookie, and we must do this by address because to do this by name would require us to export `_except_handler4`.