Merged PR 480776: /INCREMENTAL:NO support (intercepting small functions)
MSVC PR: !474074

In order to support weak functions for ~~/MD~~ all configurations post-one-dll changes, we use the interception code in order to replace the default weak implementations in the ASAN DLL with the one supplied by the user program. Originally, our interceptor declined overriding a function if there was an early `ret`, so there were a small number of functions that I just added a small nop-slide to during the one-dll changes to allow weak function registration to work. After passing `/profile` for compliance, this was insufficient - `/profile` implies `/incremental:no` which caused a large number of weak functions in libfuzzer to stop creating ILT entries where we were previously able to apply a simple jump redirect in order to intercept. Instead of injecting nop-slides everywhere, I modified the interception code to be able to read instructions in these small default implementations of the weak functions, notably `ret`. `ret` was previously not permitted to be intercepted because it is a control flow operation (which generally require a more intricate instruction decoder to be able to move the instructions into a trampoline properly), but `ret` operates correctly in that context, so I have moved it to be permitted.

Related work items: #1742408, #1742409