Merged PR 490336: CI Fix with python package version number
This fix is necessary for our CI. For some reason, the python warning gets interpreted as an error when executed through the workflow scripts. This just changes the version number of the package to eliminate the warning.

In the future, if possible, we should use pip for the test framework.