Merged PR 491323: [ASan] fix unit tests
See [MSVC-PR-490940][] and [LLVM-D158221][].

It turns out that the STL requires you to place the DLLs into the
directory of the executable; else, `msvcp140.dll` will be taken from
`C:\Windows\System32` (which is unsupported).

This was discovered because of the recent constexpr mutex change;
see [microsoft/STL#3824][]. Especially the fuzzer unit tests fail
completely with the mutex changes.

[MSVC-PR-490940]: https://dev.azure.com/devdiv/DevDiv/_git/msvc/pullrequest/490940
[LLVM-D158221]: https://reviews.llvm.org/D158221
[microsoft/STL#3824]: https://github.com/microsoft/STL/issues/3824