Merged PR 493407: Fixes for global/local friends
**MSVC PR**: [MSVC-PR-493408]

This is the same PR as @<Zachary Johnson>'s [ASan-PR-475406][],, since I can't push to that branch.

[MSVC-PR-493408]: https://devdiv.visualstudio.com/DevDiv/_git/msvc/pullrequest/493408
[ASan-PR-475406]: https://devdiv.visualstudio.com/DevDiv/_git/AddressSanitizer/pullrequest/475406

Whenever Global* or Local* functions are used prior to ASAN initialization, after ASAN is initialized, we will properly categorize the allocation as allocated prior to ASAN's initialization. The logic is mainly in asan_malloc_win.cpp and some in asan_malloc_win_moveable.cpp. CheckGlobalLocalHeapOwnership and NotOwnedByASAN are the two functions to focus on.

There is a rewrite of the tests we had previously regarding the "outer space" memory. My hope is that we can better utilize it now to accommodate different heap flags or types of memory as the issues either arise or we have the bandwidth to address them.

Lastly, there is also a fix for creating the main thread before we set asan_init=1. XStore is still testing that one, but I believe moving that before we set asan_init=1 is still valid.