[ASan][Windows] Fix rip-relative instruction replacement
The old code incorrectly checked what relative offsets were allowed.

The correct check is that the offset from the target to the instruction
pointer should be within $[-2^{31}, 2^{31})$; however, the check that was
originally written was that the offset was within $[0, 2^{31})$.
Negative offsets are certainly allowable (as long as they fit in 32 bits),
and this change fixes that.

(cherry picked from commit f855b36b482aaf8b836ebf87b1ddf83757b70ca5)