Merged PR 522532: Fixes for COE regarding environments and outputs
**MSVC PR:** !522533

If standard output / standard error isn't available, we would crash on COE initialization (this happens in a WinMain application).

If `__asan_default_options` was used rather than setting `ASAN_OPTIONS` environment variable, we would not initialize to the correct values since the weak functions are called after initialization. I added a `TODO` to further check values of flags and check functions that may need to be called again as a result of the weak hooks.

Related work items: #1922602