Merged PR 528175: Avoid calling `__asan_handle_no_return` for debug print exceptions
This is a pretty hacky fix to work around the bug causing the ASAN shadow to get cleared out during report generation.

`OutputDebugStringA` can get called while generating a report, which in turn raises an exception that does not alter control flow.
If `__asan_handle_no_return` is called, the entire stack shadow will be cleared, causing a report to falsely show that the stack contents are accessible.

Related work items: #1955253, #1969223