Merged PR 530809: Updates to readme
Changing the structure of the READMEs a bit to better reflect the state of the fork as well as how to get started. It's best viewed by viewing the files with "preview" instead of "raw content".

We can't upstream this, but I think it's relevant for anyone trying to develop with us.