Merged PR 535987: Fix for long jumps greater than 2GB away bug
I removed the first change, which was creating a long jump by moving the target into rax and jumping to it.

This is now just a fix for integer overflow. The result of this operation is outside the range of int, which leads to undefined behavior, but is _usually_ correct.

Related work items: #1929297