Merged PR 531576: Fix to output COE to console
**MSVC PR:** !533148

Related to #1957991

This addresses the issue of the console output not appearing for `WinMain` applications. If the user decides to allocate `stdout` or `stderr`, we need to update the `coe_res_file_handle` as the weak callback for `__asan_default_options` will have already been called.

In the previous fix !522532 , we just created a default log file named `asan_coe.log` without respecting whether or not the user may allocate a console. I removed this block, as now we should always respect it.

Also, there can be an issue where ASan allocators have already been initialized and allocations take place, but COE is not yet initialized. We need to either retroactively track them, or ignore them. I chose the former for the time being.