Merged PR 525894: Fix failing test deadlock_on_process_shutdown.cpp
MSVC PR: !525805

The test `deadlock_on_process_shutdown.cpp` spawns a process and waits for it to terminate. If the terminate doesn't happen after waiting for a while, we assume that a deadlock has occurred.
This test has been flaky for a while now, with the timeout limit being reached before the process terminates. Increasing the timeout fixes this, allowing the process to terminate naturally, hinting that the previous timeout value was simply too small.
Note: Invoking the test exe _directly_ never reproduced a failure; it's only when the invocation was via the python test environment (i.e., python's `thread.start()` ) that the timeout occurred, further indicating that the issue was simply one of slowness.