Merged PR 534702: Dynamically handle memcpy/memmove aliasing
MSVC PR: !538340

Our implementation previously had hard-coded assumptions about when memcpy is an alias for memmove, but Windows Insider changes have broken these (specifically, funcoverrides are now implemented for memmove, in which case memcpy is *not* an alias).
This PR removes the need for hard-coded alias assumptions to guide which functions to intercept and which to skip. Instead, we now track which addresses have been intercepted, and simply avoid re-intercepting the same address twice.