Merged PR 541670: COE fixes
**MSVC PR:** !541669

COE can now handle subprocesses launching by creating files of the same name with PID appended rather than crashing. Also fixed some printing spaces in error output.

Currently there are some places in the code  that have `flags()->continue_on_error = true` which is meant to be either `0` `1` or `2` based on _where_ you want your COE output to go. It's a little confusing to me that you can also define the `COE_LOG_FILE` instead. I also added the check for the existence to that in the weak callback flags function so we check if the environment variable was programmatically set and turn on COE.

I've adapted  Jim's test harness to be `coe_stress_test.cpp`.

!548187 is a follow up that addresses various comments below

Related work items: #1993485