Merged PR 552121: Revert breakage to unit test object file naming
In my previous PR !546502 I had deleted the *entire* block as a driveby cleanup:
```
#elsewhere asan_lib and asan_cxx_lib should still resolve to the regular library names
__testConfig.substitutions |= {
    out_to_obj_tuple,
    out_to_exe_tuple,
    ("%asan_lib", __litConfig.compiler_rt_libdir + "\\" + import_lib + ""),
    ("%asan_cxx_lib", __litConfig.compiler_rt_libdir + "\\" + runtime_thunk + ""),
}
```
The intent was to remove the unused `asan_lib` and `asan_cxx_lib` substitutions in that branch (none of our tests use those two variables except `unsymbolized.cpp`, which is covered in a preceeding branch). However,
the substitutions `out_to_obj_tuple` and `out_to_exe_tuple` turn out to be necessary, and caused a single fuzzer test to fail (which didn't raise a big flag because our fuzzer tests are known to be flaky).

This PR reverts the change to that whole block just to be safe.