Merged PR 562795: Print out any exceptions from LLVMFuzzerInitialize
In OneFuzz we are seeing people get confused by failures here.

It's becoming more and more common for OS code to throw things like `wil::exception` and when this happens inside `LLVMFuzzerInitialize` it follows the normal unhandled-exception sequence where `std::terminate` is called, then `std::abort` which ends up doing a fail-fast, and the process finally exits with `0xc0000409` (`STATUS_STACK_BUFFER_OVERRUN`).

This is particularly confusing in the context of a fuzzer because the exit code makes it look like ASan detected a legitimate memory issue, when the actual problem is only an unhandled exception.

This change prints out the exception message to make it clear what is happening, but it does not change the consequence to an `exit(1)` because I think it's still a good thing to participate in the normal error-handling flow for WER and so on.