Merged PR 569298: Adding MEM_NODISCARD to supported flags for legacy allocations
**MSVC PR:** !569915

Since `*MEM_NODISCARD` was not a part of the supported allocation flags list, calls to allocate from the legacy allocators with that flag would go to the REAL functions. Behavior is particularly interesting when mixing an unsupported flag with 0 as size of the allocation.

From the documentation:
> If this parameter is zero and the uFlags parameter specifies `GMEM_MOVEABLE`, the function returns a handle to a memory object that is marked as discarded.

If this combination occurs currently, whenever we are checking `NotOwnedByASAN`, we will return fail to look up the handle and return `OWNED_BY_UNKNOWN`. The `REAL` allocation returned will be just a handle from somewhere on a zero'd page of memory for handles from `RtlAllocateHandle`.

If a size other than `0` is passed in, [that handle has some meaningful information stored in it and can be dereferenced](https://microsoft.visualstudio.com/OS/_git/os.2020?path=/minkernel/kernelbase/gmem.c&version=GBofficial/main&line=84&lineEnd=95&lineStartColumn=1&lineEndColumn=14&lineStyle=plain&_a=contents). This means our `IsHandleOnProcessHeap` check succeeds, and we defer back to `REAL` functions.

If a size of `0` is passed in, the handle information does not need to be updated and remains null, which means our checks later on will fail. If we had a way to look into the `BaseHeapHandleTable->CommittedHandles` from the [legacy allocators](https://microsoft.visualstudio.com/OS/_git/os.2020?path=/minkernel/kernelbase/kernelbase.h&version=GBofficial/main&line=129&lineEnd=130&lineStartColumn=1&lineEndColumn=1&lineStyle=plain&_a=contents), we could check there for the existence of a committed handle of size `0`. We can't know yet if we have a wild pointer, and it's unsafe to pass that to the `*Lock` functions to check and make sure we have defined memory.

All of this to say, [that flag is obsolete](https://learn.microsoft.com/en-us/windows/win32/api/winbase/nf-winbase-globalalloc#:~:text=The%20following%20values%20are,GMEM_SHARE). The only references I could find to it are related to 16-bit Windows compatibility. I think adding it to the supported flags is the best path forward.

Related work items: #1943912