Merged PR 567133: Remove CRT init dependency
MSVC PR: !553424

Make the global map `AddressInfoMap` use the `immortalize` pattern, to ensure that it is initialized prior to first use even if the `CRT` has not yet been initialized. This is needed for instrumenting vcruntime with ASan, because the CRT isn't being used at the point when ASan is invoked.