Merged PR 575431: Handling invalid addresses for error report callbacks
**MSVC PR:** !575430

If there is an error report callback that exists in a DLL that is unloaded, we will end up crashing while trying to execute it during an error report.

This also affects applications that do not have ASan instrumentation but load a library that has ASan instrumentation and links to `vcasan.lib` -- this was the original case that caught the issue.

We discussed it once, but now with the additional `continue_on_error` context, I'm wondering if we need to wrap the entirety of the reporting machinery in structured exception handling to prevent the user from losing a valid report.

There's also a small bug fix for using `internal_memcpy` in COE.

Related work items: #2235052